+++ CHANGES.md
@@ -22,6 +22,7 @@
 -   Use pseudo url for data file stored in storage api #3000
 -   Turn off `Open In National Map` Button for Internal Storage data file Link
 -   Use [magda-preview-map](https://github.com/magda-io/magda-preview-map) v0.0.58
+-   Fixed: Dataset Creation Tool may not always capture format properly #3001
 
 ## 0.0.57